update barang sewa

+++ app/Repository/Admin/UserRepository.php
@@ -146,6 +146,9 @@ class UserRepository
     public function delete(Request $user)
     {
         try {
+            if ($user->id == 1) {
+                return response()->json(false, 500);
+            }
             $user = User::find($user->id);
             $user->delete();
             return response()->json();
